# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pulse Messenger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chats:
    get:
      operationId: getChats
      tags: [chats]
      summary: Get all chats for current user
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chat"
    post:
      operationId: createChat
      tags: [chats]
      summary: Create a new chat or group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChatBody"
      responses:
        "201":
          description: Created chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{chatId}:
    get:
      operationId: getChatById
      tags: [chats]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    put:
      operationId: updateChat
      tags: [chats]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChatBody"
      responses:
        "200":
          description: Updated chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
    delete:
      operationId: deleteChat
      tags: [chats]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Chat deleted
  /chats/{chatId}/members:
    get:
      operationId: getChatMembers
      tags: [chats]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMember"
    post:
      operationId: addChatMember
      tags: [chats]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemberBody"
      responses:
        "201":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMember"
  /chats/{chatId}/members/{memberId}:
    delete:
      operationId: removeChatMember
      tags: [chats]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Member removed
  /chats/{chatId}/pin:
    put:
      operationId: pinChat
      tags: [chats]
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat pinned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
...
  /chats/{chatId}/read:
    post:
      operationId: markChatAsRead
      tags: [chats]
      summary: Mark all messages in a chat as read
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Marked as read
  /posts:
    get:
      operationId: getPosts
      tags: [posts]
      summary: Get feed posts
      responses:
        "200":
          description: List of posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a new post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostBody"
      responses:
        "201":
          description: Post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}:
    delete:
      operationId: deletePost
      tags: [posts]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Post deleted
  /posts/{postId}/like:
    post:
      operationId: likePost
      tags: [posts]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked/unliked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Post"
  /posts/{postId}/comments:
    get:
      operationId: getPostComments
      tags: [posts]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostComment"
    post:
      operationId: createPostComment
      tags: [posts]
      parameters:
        - name: postId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostCommentBody"
      responses:
        "201":
          description: Comment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostComment"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        bio:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        avatarColor:
          type: string
        status:
          type: string
          enum: [online, offline, away]
        statusText:
          type: string
          nullable: true
        lastSeen:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        isBot:
          type: boolean
        isVerified:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - username
        - displayName
        - avatarColor
        - status
        - isBot
        - isVerified
        - createdAt
...
    Chat:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [direct, group, channel]
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        avatarColor:
          type: string
        isPinned:
          type: boolean
        isMuted:
          type: boolean
        unreadCount:
          type: integer
        lastMessage:
          $ref: "#/components/schemas/Message"
          nullable: true
        members:
          type: array
          items:
            $ref: "#/components/schemas/ChatMember"
        otherUser:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - type
        - avatarColor
        - isPinned
        - isMuted
        - unreadCount
        - createdAt
        - updatedAt
    CreateChatBody:
      type: object
      properties:
        type:
          type: string
          enum: [direct, group, channel]
        name:
          type: string
        description:
          type: string
        memberIds:
          type: array
          items:
            type: integer
      required:
        - type
    UpdateChatBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        avatarUrl:
          type: string
        isMuted:
          type: boolean
    ChatMember:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        userId:
          type: integer
        role:
          type: string
          enum: [owner, admin, member]
        joinedAt:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - id
        - chatId
        - userId
        - role
        - joinedAt
    AddMemberBody:
      type: object
      properties:
        userId:
          type: integer
        role:
          type: string
          enum: [admin, member]
          default: member
      required:
        - userId
    Message:
      type: object
      properties:
        id:
          type: integer
        chatId:
          type: integer
        senderId:
          type: integer
        text:
          type: string
          nullable: true
        type:
          type: string
          enum: [text, image, video, audio, file, gift, call, sticker, system]
        mediaUrl:
          type: string
          nullable: true
        replyToId:
          type: integer
          nullable: true
        isEdited:
          type: boolean
        isRead:
          type: boolean
        reactions:
          type: array
          items:
            $ref: "#/components/schemas/Reaction"
        sender:
          $ref: "#/components/schemas/User"
          nullable: true
        replyTo:
          $ref: "#/components/schemas/Message"
          nullable: true
        giftData:
          $ref: "#/components/schemas/Gift"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - chatId
        - senderId
        - type
        - isEdited
        - isRead
        - reactions
        - createdAt
        - updatedAt
...
    Reaction:
      type: object
      properties:
        id:
          type: integer
        messageId:
          type: integer
        userId:
          type: integer
        emoji:
          type: string
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - messageId
        - userId
        - emoji
        - createdAt
...
    GiftItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        emoji:
          type: string
        animationType:
          type: string
          enum: [confetti, sparkle, hearts, fireworks, stars, balloons, roses, diamonds]
        rarity:
          type: string
          enum: [common, rare, epic, legendary]
        stars:
          type: integer
        description:
          type: string
      required:
        - id
        - name
        - emoji
        - animationType
        - rarity
        - stars
        - description
    Gift:
      type: object
      properties:
        id:
          type: integer
        giftItemId:
          type: integer
        senderId:
          type: integer
        receiverId:
          type: integer
        message:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        giftItem:
          $ref: "#/components/schemas/GiftItem"
          nullable: true
        sender:
          $ref: "#/components/schemas/User"
          nullable: true
        receiver:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - giftItemId
        - senderId
        - receiverId
        - isAnonymous
        - createdAt
...
    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        text:
          type: string
        imageUrl:
          type: string
          nullable: true
        likesCount:
          type: integer
        commentsCount:
          type: integer
        isLiked:
          type: boolean
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - userId
        - text
        - likesCount
        - commentsCount
        - isLiked
        - createdAt
    PostComment:
      type: object
      properties:
        id:
          type: integer
        postId:
          type: integer
        userId:
          type: integer
        text:
          type: string
        author:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - postId
        - userId
        - text
        - createdAt
    CreatePostBody:
      type: object
      properties:
        text:
          type: string
        imageUrl:
          type: string
      required:
        - text
    CreatePostCommentBody:
      type: object
      properties:
        text:
          type: string
      required:
        - text
...